+++ SourceGithub/SourceGithub.php
@@ -11,8 +11,9 @@ require_once( config_get( 'core_path' ) . 'json_api.php' );
 
 class SourceGithubPlugin extends MantisSourceGitBasePlugin {
 
-	const PLUGIN_VERSION = '2.0.2';
-	const FRAMEWORK_VERSION_REQUIRED = '2.0.0';
+	const PLUGIN_VERSION = '2.1.0';
+	const FRAMEWORK_VERSION_REQUIRED = '2.2.0';
+	const MANTIS_VERSION = '2.3.0';
 
 	public $linkPullRequest = '/pull/%s';
 
@@ -32,6 +33,161 @@ class SourceGithubPlugin extends MantisSourceGitBasePlugin {
 	}
 
 	public $type = 'github';
+
+	public function resources( $p_event ) {
+		# Only include the javascript when it's actually needed
+		parse_str( parse_url( $_SERVER['REQUEST_URI'], PHP_URL_QUERY ), $t_query );
+		if( array_key_exists( 'page', $t_query ) ) {
+			$t_page = basename( $t_query['page'] );
+			if( $t_page == 'repo_update_page' ) {
+				return '<script src="' . plugin_file( 'sourcegithub.js' ) . '"></script>';
+			}
+		}
+		return null;
+	}
+
+	public function hooks() {
+		return parent::hooks() + array(
+			"EVENT_LAYOUT_RESOURCES" => "resources",
+			'EVENT_REST_API_ROUTES' => 'routes',
+		);
+	}
+
+	/**
+	 * Add the RESTful routes handled by this plugin.
+	 *
+	 * @param string $p_event_name The event name
+	 * @param array  $p_event_args The event arguments
+	 * @return void
+	 */
+	public function routes( $p_event_name, $p_event_args ) {
+		$t_app = $p_event_args['app'];
+		$t_plugin = $this;
+		$t_app->group(
+			plugin_route_group(),
+			function() use ( $t_app, $t_plugin ) {
+				$t_app->delete( '/{id}/token', [$t_plugin, 'route_token_revoke'] );
+				$t_app->post( '/{id}/webhook', [$t_plugin, 'route_webhook'] );
+			}
+		);
+	}
+
+	/**
+	 * RESTful route to revoke GitHub application token
+	 *
+	 * @param Slim\Http\Request  $p_request
+	 * @param Slim\Http\Response $p_response
+	 * @param array              $p_args
+	 *
+	 * @return Slim\Http\Response
+	 */
+	public function route_token_revoke( $p_request, $p_response, $p_args ) {
+		# Make sure the given repository exists
+		$t_repo_id = isset( $p_args['id'] ) ? $p_args['id'] : $p_request->getParam( 'id' );
+		if( !SourceRepo::exists( $t_repo_id ) ) {
+			return $p_response->withStatus( HTTP_STATUS_BAD_REQUEST, 'Invalid Repository Id' );
+		}
+
+		# Check that the repo is of GitHub type
+		$t_repo = SourceRepo::load( $t_repo_id );
+		if( $t_repo->type != $this->type ) {
+			return $p_response->withStatus( HTTP_STATUS_BAD_REQUEST, "Id $t_repo_id is not a GitHub repository" );
+		}
+
+		# Clear the token
+		unset( $t_repo->info['hub_app_access_token'] );
+		$t_repo->save();
+
+		return $p_response->withStatus( HTTP_STATUS_NO_CONTENT );
+	}
+
+	/**
+	 * RESTful route to create GitHub checkin webhook
+	 *
+	 * @param Slim\Http\Request  $p_request
+	 * @param Slim\Http\Response $p_response
+	 * @param array              $p_args
+	 *
+	 * @return Slim\Http\Response
+	 */
+	public function route_webhook( $p_request, $p_response, $p_args ) {
+		plugin_push_current( 'Source' );
+
+		# Make sure the given repository exists
+		$t_repo_id = isset( $p_args['id'] ) ? $p_args['id'] : $p_request->getParam( 'id' );
+		if( !SourceRepo::exists( $t_repo_id ) ) {
+			return $p_response->withStatus( HTTP_STATUS_BAD_REQUEST, 'Invalid Repository Id' );
+		}
+
+		# Check that the repo is of GitHub type
+		$t_repo = SourceRepo::load( $t_repo_id );
+		if( $t_repo->type != $this->type ) {
+			return $p_response->withStatus( HTTP_STATUS_BAD_REQUEST, "Id $t_repo_id is not a GitHub repository" );
+		}
+
+		$t_username = $t_repo->info['hub_username'];
+		$t_reponame = $t_repo->info['hub_reponame'];
+
+		# GitHub webhook payload URL
+		$t_payload_url = config_get( 'path' ) . plugin_page( 'checkin', true )
+			. '&api_key=' . plugin_config_get( 'api_key' );
+
+		# Retrieve existing webhooks
+		try {
+			$t_github_api = new \GuzzleHttp\Client();
+			$t_api_uri = SourceGithubPlugin::api_uri( $t_repo, "repos/$t_username/$t_reponame/hooks" );
+
+			$t_response = $t_github_api->get( $t_api_uri );
+		}
+		catch( GuzzleHttp\Exception\ClientException $e ) {
+			return $e->getResponse();
+		}
+		$t_hooks = json_decode( (string) $t_response->getBody() );
+
+		# Determine if there is already a webhook attached to the plugin's payload URL
+		$t_id = false;
+		foreach( $t_hooks as $t_hook ) {
+			if(   $t_hook->name == 'web' && $t_hook->config->url == $t_payload_url ) {
+				$t_id = $t_hook->id;
+				break;
+			}
+		}
+
+		if( $t_id ) {
+			# Webhook already exists for this URL
+			# Set the Github URL so user can easily link to it
+			$t_hook->web_url = "https://github.com/$t_username/$t_reponame/settings/hooks/" . $t_hook->id;
+			return $p_response
+				->withStatus( HTTP_STATUS_CONFLICT,
+					plugin_lang_get( 'webhook_exists', 'SourceGithub' ) )
+				->withJson( $t_hook );
+		}
+
+		# Create new webhook
+		try {
+			$t_payload = array(
+				'name' => 'web',
+				'config' => array(
+					'url' => $t_payload_url,
+					'content_type' => 'json',
+					'secret' => $t_repo->info['hub_webhook_secret'],
+				)
+			);
+
+			$t_github_response = $t_github_api->post( $t_api_uri,
+				array( GuzzleHttp\RequestOptions::JSON => $t_payload )
+			);
+		}
+		catch( GuzzleHttp\Exception\ClientException $e ) {
+			return $e->getResponse();
+		}
+
+		return $p_response
+			->withStatus( HTTP_STATUS_CREATED,
+				plugin_lang_get( 'webhook_success', 'SourceGithub' ) )
+			->withHeader('Content-type', 'application/json')
+			->write( $t_github_response->getBody() );
+	}
 
 	public function show_type() {
 		return plugin_lang_get( 'github' );
@@ -95,6 +251,7 @@ class SourceGithubPlugin extends MantisSourceGitBasePlugin {
 		$t_hub_app_client_id = null;
 		$t_hub_app_secret = null;
 		$t_hub_app_access_token = null;
+		$t_hub_webhook_secret = null;
 
 		if ( isset( $p_repo->info['hub_username'] ) ) {
 			$t_hub_username = $p_repo->info['hub_username'];
@@ -114,6 +271,10 @@ class SourceGithubPlugin extends MantisSourceGitBasePlugin {
 
 		if ( isset( $p_repo->info['hub_app_access_token'] ) ) {
 			$t_hub_app_access_token = $p_repo->info['hub_app_access_token'];
+		}
+
+		if ( isset( $p_repo->info['hub_webhook_secret'] ) ) {
+			$t_hub_webhook_secret = $p_repo->info['hub_webhook_secret'];
 		}
 
 		if ( isset( $p_repo->info['master_branch'] ) ) {
@@ -141,33 +302,73 @@ class SourceGithubPlugin extends MantisSourceGitBasePlugin {
 <tr>
 	<td class="category"><?php echo plugin_lang_get( 'hub_app_client_id' ) ?></td>
 	<td>
-		<input type="text" name="hub_app_client_id" maxlength="250" size="40" value="<?php echo string_attribute( $t_hub_app_client_id ) ?>"/>
+		<input name="hub_app_client_id" id="hub_app_client_id"
+			   type="text" maxlength="250" size="40"
+			   value="<?php echo string_attribute( $t_hub_app_client_id ) ?>"
+			   data-original="<?php echo string_attribute( $t_hub_app_client_id ) ?>"
+		/>
 	</td>
 </tr>
 
 <tr>
 	<td class="category"><?php echo plugin_lang_get( 'hub_app_secret' ) ?></td>
 	<td>
-		<input type="text" name="hub_app_secret" maxlength="250" size="40" value="<?php echo string_attribute( $t_hub_app_secret ) ?>"/>
+		<input name="hub_app_secret" id="hub_app_secret"
+			   type="text" maxlength="250" size="40"
+			   value="<?php echo string_attribute( $t_hub_app_secret ) ?>"
+			   data-original="<?php echo string_attribute( $t_hub_app_secret ) ?>"
+		/>
 	</td>
 </tr>
 
 <tr>
 	<td class="category"><?php echo plugin_lang_get( 'hub_app_access_token' ) ?></td>
 	<td>
-		<?php
-		if( empty( $t_hub_app_client_id ) || empty( $t_hub_app_secret ) ) {
-			echo plugin_lang_get( 'hub_app_client_id_secret_missing' );
-		} elseif( empty( $t_hub_app_access_token ) ) {
-			print_link( $this->oauth_authorize_uri( $p_repo ), plugin_lang_get( 'hub_app_authorize' ) );
-		} else {
-			echo plugin_lang_get( 'hub_app_authorized' );
-			# @TODO This would be better with an AJAX, but this will do for now
-			?>
-			<input type="submit" class="btn btn-xs btn-primary btn-white btn-round" name="revoke" value="<?php echo plugin_lang_get( 'hub_app_revoke' ) ?>"/>
+		<div id="id_secret_missing" class="hidden">
+			<?php echo plugin_lang_get( 'hub_app_client_id_secret_missing' ); ?>
+		</div>
+
+		<div id="token_missing" class="sourcegithub_token hidden">
 			<?php
-		}
-		?>
+			print_small_button(
+				$this->oauth_authorize_uri( $p_repo ),
+				plugin_lang_get( 'hub_app_authorize' )
+			);
+			?>
+		</div>
+
+		<div id="token_authorized" class="sourcegithub_token hidden">
+			<input name="hub_app_access_token" id="hub_app_access_token"
+				   type="hidden" maxlength="250" size="40"
+				   value="<?php echo string_attribute( $t_hub_app_access_token ) ?>"
+			/>
+			<?php echo plugin_lang_get( 'hub_app_authorized' ); ?>&nbsp;
+			<button id="btn_auth_revoke" type="button"
+					class="btn btn-primary btn-white btn-round btn-sm"
+					data-token-set="<?php echo $t_hub_app_access_token ? 'true' : 'false' ?>"
+			>
+				<?php echo plugin_lang_get( 'hub_app_revoke' ) ?>
+			</button>
+		</div>
+
+	</td>
+</tr>
+
+<tr>
+	<td class="category"><?php echo plugin_lang_get( 'hub_webhook_secret' ) ?></td>
+	<td>
+		<input type="text" name="hub_webhook_secret" maxlength="250" size="40" value="<?php echo string_attribute( $t_hub_webhook_secret ) ?>"/>
+		<div id="webhook_create" class="sourcegithub_token hidden">
+			<div class="space-2"></div>
+			<button type="button" class="btn btn-primary btn-white btn-round btn-sm">
+				<?php echo plugin_lang_get( 'webhook_create' ); ?>
+			</button>
+
+			<span id="webhook_status">
+				<i class="ace-icon fa fa-lg"></i>
+				<span></span>
+			</span>
+		</div>
 	</td>
 </tr>
 
@@ -181,17 +382,19 @@ class SourceGithubPlugin extends MantisSourceGitBasePlugin {
 	}
 
 	public function update_repo( $p_repo ) {
-		# Revoking previously authorized Github access token
-		if( gpc_isset( 'revoke' ) ) {
-			unset( $p_repo->info['hub_app_access_token'] );
-			return $p_repo;
-		}
-
 		$f_hub_username = gpc_get_string( 'hub_username' );
 		$f_hub_reponame = gpc_get_string( 'hub_reponame' );
 		$f_hub_app_client_id = gpc_get_string( 'hub_app_client_id' );
 		$f_hub_app_secret = gpc_get_string( 'hub_app_secret' );
+		$f_hub_webhook_secret = gpc_get_string( 'hub_webhook_secret' );
 		$f_master_branch = gpc_get_string( 'master_branch' );
+
+		# Clear the access token if client id and secret changed
+		if(   $p_repo->info['hub_app_client_id'] != $f_hub_app_client_id
+			|| $p_repo->info['hub_app_secret'] != $f_hub_app_secret
+		) {
+			unset($p_repo->info['hub_app_access_token']);
+		}
 
 		$this->validate_branch_list( $f_master_branch );
 
@@ -199,6 +402,7 @@ class SourceGithubPlugin extends MantisSourceGitBasePlugin {
 		$p_repo->info['hub_reponame'] = $f_hub_reponame;
 		$p_repo->info['hub_app_client_id'] = $f_hub_app_client_id;
 		$p_repo->info['hub_app_secret'] = $f_hub_app_secret;
+		$p_repo->info['hub_webhook_secret'] = $f_hub_webhook_secret;
 		$p_repo->info['master_branch'] = $f_master_branch;
 
 		return $p_repo;
@@ -242,9 +446,14 @@ class SourceGithubPlugin extends MantisSourceGitBasePlugin {
 	}
 
 	public function precommit() {
+		# Legacy GitHub Service sends the payload via eponymous form variable
 		$f_payload = gpc_get_string( 'payload', null );
 		if ( is_null( $f_payload ) ) {
-			return;
+			# If empty, retrieve the webhook's payload from the body
+			$f_payload = file_get_contents( 'php://input' );
+			if ( is_null( $f_payload ) ) {
+				return;
+			}
 		}
 
 		if ( false === stripos( $f_payload, 'github.com' ) ) {
@@ -268,6 +477,29 @@ class SourceGithubPlugin extends MantisSourceGitBasePlugin {
 			$t_repo->id = $t_row['id'];
 
 			if ( $t_repo->info['hub_reponame'] == $t_reponame ) {
+				# Retrieve the payload's signature from the request headers
+				# Reference https://developer.github.com/webhooks/#delivery-headers
+				$t_signature = null;
+				if( array_key_exists( 'HTTP_X_HUB_SIGNATURE', $_SERVER ) ) {
+					$t_signature = explode( '=', $_SERVER['HTTP_X_HUB_SIGNATURE'] );
+					if( $t_signature[0] != 'sha1' ) {
+						# Invalid hash - as per docs, only sha1 is supported
+						return;
+					}
+					$t_signature = $t_signature[1];
+				}
+
+				# Validate payload against webhook secret: checks OK if
+				# - Webhook secret not defined and no signature received from GitHub, OR
+				# - Payload's SHA1 hash salted with Webhook secret matches signature
+				$t_secret = $t_repo->info['hub_webhook_secret'];
+				$t_valid = ( !$t_secret && !$t_signature )
+					|| $t_signature == hash_hmac('sha1', $f_payload, $t_secret);
+				if( !$t_valid ) {
+					# Invalid signature
+					return;
+				}
+
 				return array( 'repo' => $t_repo, 'data' => $t_data );
 			}
 		}
@@ -458,7 +690,16 @@ class SourceGithubPlugin extends MantisSourceGitBasePlugin {
 		}
 
 		if ( !empty( $t_hub_app_client_id ) && !empty( $t_hub_app_secret ) ) {
-			return 'https://github.com/login/oauth/authorize?client_id=' . $t_hub_app_client_id . '&redirect_uri=' . urlencode(config_get('path') . 'plugin.php?page=SourceGithub/oauth_authorize&id=' . $p_repo->id ) . '&scope=repo';
+			$t_redirect_uri = config_get( 'path' )
+				. plugin_page( 'oauth_authorize', true ) . '&'
+				. http_build_query( array( 'id' => $p_repo->id ) );
+			$t_param = array(
+				'client_id' => $t_hub_app_client_id,
+				'redirect_uri' => $t_redirect_uri,
+				'scope' => 'repo',
+				'allow_signup' => false,
+			);
+			return 'https://github.com/login/oauth/authorize?' . http_build_query( $t_param );
 		} else {
 			return '';
 		}